Add tests for setTimeout/setInterval default delay
Add a test to ensure setTimeout's delay defaults to zero.

Fix setInterval's test for the default delay. Before, the test passed
in zero as the delay, which goes against the whole purpose of testing
whether the delay defaults to zero when nothing is passed in.

Make sure all other references to setTimout/setInterval in tests pass
in a delay so that it doesn't get defaulted.